Override Hugo sitemap and hide legacy pages on it
The 'quickstart' and 'tutorials' pages under the root /docs folder are
to be left, but hidden from in-page navigation and indexing. This uses
the example-site Docsy sitemap.xml file to override the default Hugo
one. A frontmatter param of 'private' is used as a flag to hide the page
from the sitemap.